requirements.txt
PyYAML
attrs==19.3.0
Click==7.0
importlib-metadata==1.3.0
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.1
pyfiglet==0.8.post1
pyparsing==2.4.6
pytest==5.3.2
six==1.13.0
wcwidth==0.1.7
zipp==0.6.0
